# Uses Regex instead of fancy-regex - 6x speedup