This PR realizes the wish expressed in current code to use the faster Regex. 

The text is splitted to pieces, before tokenization, according to regular expression patterns. This RP drops a lookahead part of the pattern,  the part for catching the whitespaces, and handles the whitespaces with scripting instead,  _with mathematically provable exactly same output._ 
This makes it possible to use linear-time Regex instead of fancy-regex, as Regex does not support lookahead, resulting in a 14x speedup of pattern matching. As pattern matching currently comprises 90% of the encoding runtime, the total runtime is boosted 6x. 
    
Although fancy_regex delegates to Regex, when the pattern has no special features, it is still some 10% slower in test, thus we directly use Regex.
This improvement is for pattern matching of the parts with ordinary text. Catching the special tokens is still done with fancy_regex. 
|----|----|----|----|
| 100 MB of wikitext-103 | 22138325 | 18.94s | 4.94s |
| 100 MB of Linux code | 36119543 | 30.28s | 4.59s |