@tmm1 2024-11-09
Thanks for your work on this!

I noticed this code block which sounds like it would need to change along with these regexes?

https://github.com/openai/tiktoken/blob/63527649963def8c759b0f91f2eb69a40934e468/src/lib.rs#L405-L409