@Majdoddin 2024-11-22
@tmm1  I've implement it for encode_ordinary(). Than part is for unstable encoding.
By the way, I'd really appreciate if you submit a review for this PR.